+++ Changelog.md
@@ -3,6 +3,11 @@ WhateverGreen Changelog
 #### v1.2.5
 - Added support for specifying `agpmod` in external GPU properties
 - Added fatal error on `agpmod=cfgmap` on 10.14 and newer, which had no effect since 10.13.4
+- Added `igfxcflbklt` boot argument and `enable-cfl-backlight-fix` property to fix CFL backlight
+- Added max backlight frequency override via  `max-backlight-freq` IGPU property on CFL
+- Added `framebuffer-camellia` and `framebuffer-flags` patching
+- Added `AppleBacklight` patching code (based on `AppleBacklightFixup` by `hieplpvip`)
+- Added NVIDIA HDMI enabling code for firmwares that disable it by default (thanks @Fraxul)
 
 #### v1.2.4
 - Added platform list dumping to ioreg (at IOService:/IOResources/WhateverGreen), debug build only with -igfxfbdump